ergänze Benutzeroption für Art des killens, prüfe signal variable im Handler

+++ src/main.cpp
@@ -18,29 +18,26 @@ vector<pid_t> children;
 
 
 void signalHandler(int signal) {
-  cout << signal << endl;
- 
+
+ if(signal == 15){
   if (children.size() > 0){
-    //cout << "muss noch eine ebene tiefer gehen!" << endl;
-    //cout << "ausgabe: " << children.size() << endl;
     for(size_t c = 0; c < children.size(); c++){
-      cout << "kille : " << children[c] << endl;
-      int success = kill(children[c], SIGKILL);
-      cout << success << endl;
-      
+      cout << "killíng : " << children[c] << endl;
+      kill(children[c], SIGKILL);
     }
   }
-  
-    cout << "kille nun: " << getpid() << endl;
+  cout << "killing now: " << getpid() << endl;
+  quick_exit(EXIT_SUCCESS);
     
-    quick_exit(EXIT_SUCCESS);
-    
+  } else {
+    cout << "Got bad signal!" << endl;
+  }
 }
 
 void print_process(){
 
-  cout << "I am Prozess: " << getpid() << endl;
-  cout << "Parent id is : " << getppid() << endl;
+  cout << "I am Prozess  : " << getpid() << endl;
+  cout << "Id from my is : " << getppid() << endl;
   cout << endl;
 
 }
@@ -100,6 +97,53 @@ void fork_process(int n, int m, int depth) {
   }
 }
 
+void fork_process_auto(int n, int m, int depth){
+  int i = 0;
+  if (depth == 0) {
+    while (i < n) {
+      auto pid{fork()};
+      if (pid == -1) {
+        cerr << "forking failed : " << errno << endl;
+        exit(EXIT_FAILURE);
+      }
+      if (pid == 0) {
+        cnt1 += i;
+        cout << cnt1 << endl;
+        print_process();
+        fork_process_auto(n, m, depth + 1);
+        quick_exit(EXIT_SUCCESS);
+      } else {
+        int status;
+        waitpid(pid, &status, 0);
+      }
+      i++;
+    }
+  }
+  else if (depth == 1) {
+   
+    while (i < m) {
+      auto pid{fork()};
+      if (pid == -1) {
+        cerr << "forking failed : " << errno << endl;
+        exit(EXIT_FAILURE);
+      }
+      if (pid == 0) {
+        cnt2 += i;
+        cout << cnt1 << "." << cnt2 << endl;
+        print_process();
+        quick_exit(EXIT_SUCCESS);
+      } else {
+        int status;
+        waitpid(pid, &status, 0);
+      }
+      i++;
+    }
+  } else {
+    return;
+  }
+}
+
+
 int main(int argc, char **argv) {
   signal(SIGTERM, signalHandler);
  
@@ -107,14 +151,22 @@ int main(int argc, char **argv) {
 
   int n;
   int m;
+  bool signalBool;
 
   app.add_option("children", n, "Number of children");
   app.add_option("leaves", m, "Number of leaves");
+  app.add_flag("--killsignal, --k", signalBool, "kill leaves with signal");
+
   CLI11_PARSE(app, argc, argv);
 
   cout << "main process" << endl;
   print_process();
 
-  fork_process(n, m, 0);
+  if(signalBool){
+    fork_process(n, m, 0);
+  } else {
+    fork_process_auto(2, 2, 0);
+  }
+  
 
 }
